+++ src/test/java/at/pwd/shallowred/EndgameDB/EndgameDBTest.java
@@ -123,24 +123,24 @@ class EndgameDBTest
             int expected = i-1-(i>7?1:0);
 
             int slot = MancalaBoard.getOppositeSlot(i);
-            int[] board = new int[14];
+            byte[] board = new byte[14];
             board[MancalaBoard.index(MancalaBoard.PLAYER_A,slot)] = 1;
             //set depots to random values, should not have any influence
-            board[MancalaBoard.DEPOT_A] = ThreadLocalRandom.current().nextInt(Integer.MAX_VALUE);
-            board[MancalaBoard.DEPOT_B] = ThreadLocalRandom.current().nextInt(Integer.MAX_VALUE);
+            board[MancalaBoard.DEPOT_A] = (byte)ThreadLocalRandom.current().nextInt(Integer.MAX_VALUE);
+            board[MancalaBoard.DEPOT_B] = (byte)ThreadLocalRandom.current().nextInt(Integer.MAX_VALUE);
             args.add(Arguments.of(new MancalaGame(MancalaBoard.PLAYER_A,board),expected));
 
             //same board mirrored, with other player as current player
-            int[] boardMirrored = new int[14];
+            byte[] boardMirrored = new byte[14];
             boardMirrored[MancalaBoard.index(MancalaBoard.PLAYER_B,slot)] = 1;
             //set depots to random values, should not have any influence
-            boardMirrored[MancalaBoard.DEPOT_A] = ThreadLocalRandom.current().nextInt(Integer.MAX_VALUE);
-            boardMirrored[MancalaBoard.DEPOT_B] = ThreadLocalRandom.current().nextInt(Integer.MAX_VALUE);
+            boardMirrored[MancalaBoard.DEPOT_A] = (byte)ThreadLocalRandom.current().nextInt(Integer.MAX_VALUE);
+            boardMirrored[MancalaBoard.DEPOT_B] = (byte)ThreadLocalRandom.current().nextInt(Integer.MAX_VALUE);
             args.add(Arguments.of(new MancalaGame(MancalaBoard.PLAYER_B,boardMirrored),expected));
         }
 
         //add hand calculated example
-        args.add(Arguments.of(new MancalaGame(MancalaBoard.PLAYER_A,new int[]{0,2,0,0,0,3,0,0,1,0,0,1,0,0,0}),28229));
+        args.add(Arguments.of(new MancalaGame(MancalaBoard.PLAYER_A,new byte[]{0,2,0,0,0,3,0,0,1,0,0,1,0,0,0}),28229));
 
         return args.stream();
     }
